refactor: revision refactor
