+++ packages/react-native-kit/src/components/molecules/Calendar/Calendar.tsx
@@ -5,16 +5,17 @@ import {
 } from '@tecsinapse/react-core';
 import React from 'react';
 import { Text } from '../../atoms/Text';
-import { getLocale } from '../../../utils/dateUtils';
+import { getLocale } from '../../../utils/date';
 
-export const Calendar = <T extends SelectionType>(
-  props: CalendarProps<T>
-): JSX.Element => {
+export const Calendar = <T extends SelectionType>({
+  locale,
+  ...rest
+}): JSX.Element => {
   return (
     <CalendarCore
-      {...props}
+      {...rest}
       TextComponent={Text}
-      locale={props.locale ?? getLocale()}
+      locale={locale ?? getLocale()}
     />
   );
 };
