+++ packages/react-native-kit/src/components/molecules/DatePicker/DatePicker.tsx
@@ -6,18 +6,19 @@ import {
 import React from 'react';
 import { Text } from '../../atoms/Text';
 import { Calendar } from '../Calendar';
-import { getLocale } from '../../../utils/dateUtils';
+import { getLocale } from '../../../utils/date';
 
-export const DatePicker = <T extends SelectionType>(
-  props: DatePickerProps<T>
-): JSX.Element => {
+export const DatePicker = <T extends SelectionType>({
+  locale,
+  ...rest
+}): JSX.Element => {
   return (
     <DatePickerCore
-      {...props}
+      {...rest}
       TextComponent={Text}
       CalendarComponent={Calendar}
       animationType="slide"
-      locale={props.locale ?? getLocale()}
+      locale={locale ?? getLocale()}
     />
   );
 };
